# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para realizar factorización QR de matrices rectangulares.",
        "title": "Matrix Factorization API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Autentica a un usuario y devuelve un token JWT válido por 24h.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Inicio de sesión",
                "parameters": [
                    {
                        "description": "Credenciales de acceso",
                        "name": "credentials",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login exitoso con token JWT",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Crea un nuevo usuario en el sistema con contraseña encriptada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registro de nuevo usuario",
                "parameters": [
                    {
                        "description": "Datos de registro",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor o email duplicado",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    }
                }
            }
        },
        "/matrix/factorize": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calcula las matrices Q (ortogonal) y R (triangular superior) de una matriz dada.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matrix"
                ],
                "summary": "Realiza la factorización QR",
                "parameters": [
                    {
                        "description": "Matriz original",
                        "name": "matrix",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.QRRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.QRResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_pholguinc_api-go-matrices_internal_dtos.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Data returned by the operation"
                },
                "message": {
                    "description": "Message describing the result\n@example \"Operación exitosa\"",
                    "type": "string"
                },
                "status": {
                    "description": "Status code of the response\n@example 200",
                    "type": "integer"
                }
            }
        },
        "github_com_pholguinc_api-go-matrices_internal_dtos.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {}
            }
        },
        "github_com_pholguinc_api-go-matrices_internal_dtos.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "github_com_pholguinc_api-go-matrices_internal_dtos.QRRequest": {
            "type": "object",
            "required": [
                "matrix"
            ],
            "properties": {
                "matrix": {
                    "description": "Matrix is a 2D array of numbers\n@example [[12, -51], [6, 167]]",
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        },
        "github_com_pholguinc_api-go-matrices_internal_dtos.QRResponseData": {
            "type": "object",
            "properties": {
                "q": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                },
                "r": {
                    "type": "array",
                    "items": {
                        "type": "array",
                        "items": {
                            "type": "number",
                            "format": "float64"
                        }
                    }
                }
            }
        },
        "github_com_pholguinc_api-go-matrices_internal_dtos.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Escribe 'Bearer ' seguido de tu token JWT.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}